Updated files

+++ include/plant_base.h
@@ -2,7 +2,7 @@
 class PlantBase
 {
 public:
-    virtual double getHeight() = 0;
+    virtual float getHeight() = 0;
 
     virtual void grow(int days) = 0;
 private:
